% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordword.R
\name{diffword}
\alias{diffword}
\title{Antonyms of a given word}
\usage{
diffword(word, key = NULL)
}
\arguments{
\item{word}{Word from which antonyms should be found.}

\item{key}{User API key as assigned from bighugelabs. The default, NULL,
looks for a \code{BIGHUGELABS_KEY} environment variable, which it sets
automatically for you after specifying the key once per machine.}
}
\value{
A list of antonyms for the word across different parts of speech
  (if applicable)
}
\description{
Returns antonyms for a given word using the words API from bighugelabs
}
