% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordword.R
\name{tuna}
\alias{tuna}
\title{Access the unofficial API of thesaurus.com}
\usage{
tuna(word)
}
\arguments{
\item{word}{The word of interest.}
}
\value{
A data frame with words, definitions, similiarity indices, etc.
}
\description{
This returns the results from the thesaurus.com engine (tuna.thesaurus.com).
}
